GPS and Image based campus navigation app
This work is a proof of concept of a navigation system for places where
navigation between buildings or locations is not available on popular
map apps (google,apple etc).

We propose a community system where in users will create their custom
maps and upload them to a central server.

A user take photos of landmark buildings and locations while going
between location A and location B.

We strip the GPS loc from these pics and then create a map from them in
a single form/view.

While the app is in use - the user recieves sound alerts / image alerts
on his screen as and when he/she is within (approaching) the landmark
(20mtrs).

This was a group project and several features were identified to be
worked on - Image Processing / Computer vision (automated landmark
recognition - SIFT etc) .

Alerts if the user strays ofcourse etc.